+++ livechat/configuration/client.py
@@ -1318,3 +1318,50 @@ class ConfigurationApi35(ConfigurationApiInterface):
         return self.session.post(f'{self.api_url}/create_bot',
                                  json=payload,
                                  headers=headers)
+
+    def list_channels(self,
+                      payload: dict = None,
+                      headers: dict = None) -> httpx.Response:
+        ''' Returns all existing auto access data structures.
+
+        Args:
+            payload (dict): Custom payload to be used as request's data.
+                            It overrides all other parameters provided for the method.
+            headers (dict): Custom headers to be used with session headers.
+                            They will be merged with session-level values that are set,
+                            however, these method-level parameters will not be persisted across requests.
+
+            Returns:
+                httpx.Response: The Response object from `httpx` library,
+                                which contains a server’s response to an HTTP request.
+        '''
+        if payload is None:
+            payload = prepare_payload(locals())
+        return self.session.post(f'{self.api_url}/list_channels',
+                                 json=payload,
+                                 headers=headers)
+
+    def check_product_limits(self,
+                             plan: str = None,
+                             payload: dict = None,
+                             headers: dict = None) -> httpx.Response:
+        ''' Checks product limits for plans.
+
+            Args:
+                plan (str): License plan to check limit for.
+                payload (dict): Custom payload to be used as request's data.
+                                It overrides all other parameters provided for the method.
+                headers (dict): Custom headers to be used with session headers.
+                                They will be merged with session-level values that are set,
+                                however, these method-level parameters will not be persisted across requests.
+
+            Returns:
+                httpx.Response: The Response object from `httpx` library,
+                                which contains a server’s response to an HTTP request.
+        '''
+        if payload is None:
+            payload = prepare_payload(locals())
+        return self.session.post(
+            f'{self.api_url}/check_product_limits_for_plan',
+            json=payload,
+            headers=headers)